Added a simple label update function
Adds a simple DC to object label updater. Since currently in Islandora
Object labels are tied to XML forms form field mappings, label updates
on DSID changes on batch operations are never triggered. This simple
add-on allows dc:title valued to be used as object label for a list of
PIDs